server/requirements.txt
# Web框架
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1

# 数据库
SQLAlchemy>=2.0.23
pymysql>=1.1.0

# 环境变量
python-dotenv>=1.0.0

# JWT相关
PyJWT>=2.8.0

# 密码加密
Werkzeug>=3.0.1

# CORS支持（跨域）
Flask-CORS>=4.0.0